point-and-click-continuous-deployment-with-cloud-run/requirements.txt
Flask==1.1.2
markdown2
python_frontmatter
